r3325 deviceconfig: renamed cisco.secrets.php to switch.secrets.php deviceconfig: now expect cooked handler as argument instead of hardware and software strings deviceconfig: cisco.connector becomes ios12.connector gwRetrieveDeviceConfig(): redo with exceptions renderObjectVLANSync(): ditto

+++ inc/gateways.php
@@ -275,41 +275,57 @@ function gwRecvFileFromObject ($object_id = 0, $handlername, &$output)
 	return gwRecvFile ($endpoint, $handlername, $output);
 }
 
-function gwRetrieveDeviceConfig ($object_id = 0, &$output)
+function gwRetrieveDeviceConfig ($object_id = 0)
 {
-	if ($object_id <= 0)
-		return oneLiner (160); // invalid arguments
 	$objectInfo = spotEntity ('object', $object_id);
 	$endpoints = findAllEndpoints ($object_id, $objectInfo['name']);
 	if (count ($endpoints) == 0)
-		return oneLiner (161); // endpoint not found
+		throw new RuntimeException ('endpoint not found');
 	if (count ($endpoints) > 1)
-		return oneLiner (162); // can't pick an address
+		throw new RuntimeException ('cannot pick endpoint');
 	$endpoint = str_replace (' ', '\ ', str_replace (' ', '+', $endpoints[0]));
-	$tmpfilename = tempnam ('', 'RackTables-sendfile-');
-	$hwtype = $swtype = 'unknown';
+	$tmpfilename = tempnam ('', 'RackTables-deviceconfig-');
+	$breed = '';
 	foreach (getAttrValues ($object_id) as $record)
 	{
-		if ($record['name'] == 'SW type' && strlen ($record['o_value']))
-			$swtype = str_replace (' ', '+', execGMarker ($record['o_value']));
-		if ($record['name'] == 'HW type' && strlen ($record['o_value']))
-			$hwtype = str_replace (' ', '+', execGMarker ($record['o_value']));
+		if
+		(
+			$record['name'] == 'SW type' &&
+			strlen ($record['o_value']) &&
+			preg_match ('/^Cisco IOS 12\./', execGMarker ($record['o_value']))
+		)
+		{
+			$breed = 'ios12';
+			break;
+		}
+		if
+		(
+			$record['name'] == 'HW type' &&
+			strlen ($record['o_value']) &&
+			preg_match ('/^Foundry FastIron GS /', execGMarker ($record['o_value']))
+		)
+		{
+			$breed = 'fdry5';
+			break;
+		}
 	}
+	if ($breed == '')
+		throw new RuntimeException ('cannot pick handler for this device');
 	$outputlines = queryGateway
 	(
 		'deviceconfig',
-		array ("retrieve ${endpoint} ${tmpfilename} ${hwtype} ${swtype}")
+		array ("retrieve ${endpoint} ${breed} ${tmpfilename}")
 	);
-	$output = file_get_contents ($tmpfilename);
+	$configtext = file_get_contents ($tmpfilename);
 	unlink ($tmpfilename);
 	if ($outputlines == NULL)
-		return oneLiner (163); // unknown gateway failure
+		throw new RuntimeException ('unknown gateway failure');
 	if (count ($outputlines) != 1)
-		return oneLiner (165); // protocol violation
+		throw new RuntimeException ('gateway protocol violation');
 	if (strpos ($outputlines[0], 'OK!') !== 0)
-		return oneLiner (164, array ($outputlines[0])); // gateway failure
-	// Being here means having 'OK!' in the response.
-	return oneLiner (66, array ($handlername)); // ignore provided "Ok" text
+		throw new RuntimeException ("gateway failure: ${outputlines[0]}");
+	// Being here means it was alright.
+	return $configtext;
 }
 
 ?>
